\name{ORA-package}
\alias{ORA-package}
\docType{package}
\title{
data bionics toolbox: Overrepresentation Analysis
}
\description{
Calculation functions for overrepresentation analysis and directed acyclic graph building
}
\details{
\tabular{ll}{
Package: \tab ORA\cr
Type: \tab Package\cr
Version: \tab 1.7.3\cr
Date: \tab 2021-04-26\cr
License: \tab GPL \cr
}
}
\author{
Catharina Lippmann, Michael Thrun, Alfred Ultsch

Maintainer: Sebastian Malkusch <malkusch@med.uni-frankfurt.de>
}
\references{
Lotsch, J., Ultsch, A.: Functional abstraction for large Gene Ontology taxonomies, in: Kestler, HA., Binder, H., Schmid, M., Kraus, JM. (eds): Statistical Computing 2012, Ulmer Informatik-Berichte, pp. 6, 2012 (ISSN 0939-5091).\cr
Lippmann, C., Kringel, D., Ultsch, A., Lotsch, J.:Computational functional genomics-based approaches in analgesic drug discovery and repurposing, in: Pharmacogenomics, 2018 Jun 1;19(9):783-79 <doi:10.2217/pgs-2018-0036>.
}

\keyword{ORA}
\keyword{Overrepresentation Analysis}