\name{termId}
\alias{termId}
\title{
Casts GO-term numbers to GO-term IDs.
}
\description{
Casts GO-term numbers to GO-term IDs.
}
\usage{
termId(GOtermNr)
}
\arguments{
  \item{GOtermNr}{
Vector\code{[1:n]} of GO-term numbers, e.g. \code{8150} (numeric!).
}
}
\value{
{GOtermId}{Vector\code{[1:n]} of GO-term IDs, e.g.  \code{"GO:0008150"} (string).
}}
\author{
ALU,MT,CL
}