\name{termNr}
\alias{termNr}
\title{
Casts GO-term IDs to GO-term numbers.
}
\description{
Casts GO-term IDs to GO-term numbers.
}
\usage{
termNr(GOTermId)
}
\arguments{
  \item{GOTermId}{
Vector\code{[1:n]} of GO-term IDs, e.g. \code{"GO:0008150"} (string).
}
}

\value{
\item{GOTermNr}{
Vector\code{[1:n]} of GO-term numbers, e.g. \code{8150} (numeric!).
}}

\author{
ALU
}
