\name{ORA-package}
\alias{ORA-package}
\docType{package}
\title{
data bionics toolbox: Overrepresentation Analysis
}
\description{
Shiny Interface and calculation functions for overrepresentation analysis and directed acyclic graph building
}
\details{
\tabular{ll}{
Package: \tab ORA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-07-05\cr
License: \tab GPL \cr
}
}
\author{
Catharina Lippmann, Michael Thrun, Alfred Ultsch

Maintainer: <lippmanc@mathematik.uni-marburg.de>
}
\references{
Loetsch, J., Ultsch, A.: Functional abstraction for large Gene Ontology taxonomies, in: Kestler, HA., Binder, H., Schmid, M., Kraus, JM. (eds): Statistical Computing 2012, Ulmer Informatik-Berichte, pp. 6, 2012 (ISSN 0939-5091).\cr
Loetsch, J., Doehring, A., Mogil, J.S., Arndt, T., Geisslinger, G., Ultsch, A.: Functional genomics of pain in analgesic drug development and therapy, in: Pharmacology & Therapeutics, Volume 139, Issue 1, from pp. 60 to 70, 2013.
}

\keyword{ORA}
\keyword{Overrepresentation Analysis}